add point value bar chart
